Fixed runtime issue.
