+++ build.sbt
@@ -1,8 +1,9 @@
-name := "cassandra-connector-failure-demo"
 
-version := "1.0"
+name := "cassandra-connector-success-demo"
 
-scalaVersion := "2.10.4"
+version := "1.1"
+
+scalaVersion := "2.10.5"
 
 resolvers ++= Seq(
   "Typesafe Releases" at "http://repo.typesafe.com/ typesafe/releases",
@@ -11,14 +12,29 @@ resolvers ++= Seq(
 
 resolvers += "spray" at "http://repo.spray.io/"
 
-val sparkVersion = "1.3.1"
+// won't work unless building right from master currently. release coming that supports 1.3.1
+//val sparkVersion = "1.3.1"
 
-libraryDependencies += "org.apache.spark" % "spark-streaming_2.10" % sparkVersion % "provided"
+val sparkVersion = "1.2.1"
 
-libraryDependencies += "org.apache.spark" % "spark-core_2.10" % sparkVersion % "provided"
+libraryDependencies += "org.apache.spark" %% "spark-streaming" % sparkVersion % "provided"
 
-libraryDependencies += "com.datastax.spark" % "spark-cassandra-connector_2.10" % "1.2.0"
+libraryDependencies += "org.apache.spark" %% "spark-core" % sparkVersion % "provided"
 
+libraryDependencies += "com.datastax.spark" %% "spark-cassandra-connector" % "1.2.1"
+
+run in Compile <<= Defaults.runTask(fullClasspath in Compile, mainClass in (Compile, run), runner in (Compile, run))
+
+parallelExecution in assembly := false
+
+assemblyOption in assembly ~= { _.copy(includeScala = false) }
+
+mergeStrategy in assembly <<= (mergeStrategy in assembly) {
+    (old) => {
+      case PathList("com", "google", xs @ _*) => MergeStrategy.last
+      case x => old(x)
+    }
+  }
 
 //assemblyMergeStrategy in assembly := {
 //  case PathList("META-INF", xs@_*) => MergeStrategy.discard
