+++ launch.sh
@@ -1,1 +1,1 @@
-spark-submit --master local[6] target/scala-2.10/cassandra-connector-failure-demo-assembly-1.0.jar
+spark-submit --master local[6] target/scala-2.10/cassandra-connector-success-demo-assembly-1.1.jar
